If cmake configures with extra argmuments such as
-DCMAKE_C_COMPILER=icc -DCMAKE_CXX_COMPILER=icpc

to make a use of the intel compiler, 

the corresponding PARDISO libraries are automatically called for compilation with this revision. However, even with this addition, the executable is not threaded by intel_thread but by gnu_thread in linking. Currently, I do not know how to intel-thread using cmake. A further test for the performance improvement because of the thread difference would be interesting with the Buckley-Leverret problem regared in [5866]. Unfortunately, the BL benchmark does not work after certain versions. It is also notified to Bastian Graupner, who wants to test the benchmark for his own research purpose.

Overall, it seems that the intel compiler overshadows GNU compilers particularly on the intel multi-core or multi-socket machines. It's more expensive than AMD. Personally, I think it worths money. 




git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@6028 3895a583-e809-0410-9bb6-aa9463026377